重大修改 - 學分數支援小數點
Conflicts:
	定期評量成績單/Program.cs
	定期評量成績單/Properties/AssemblyInfo.cs
	定期評量成績單/bin/Debug/Test.dll
	定期評量成績單/bin/Debug/Test.pdb
	定期評量成績單/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	定期評量成績單/obj/x86/Debug/Test.dll
	定期評量成績單/obj/x86/Debug/Test.pdb
	定期評量成績單/obj/x86/Debug/定期評量成績單.csproj.FileListAbsolute.txt

+++ SmartSchool.Evaluation.HS/Configuration/ScoreCalcRuleEditor.cs
@@ -1117,8 +1117,8 @@ namespace SmartSchool.Evaluation.Configuration
                 }
                 else
                 {
-                    int i;
-                    if (int.TryParse(txtCredit.Text, out i))
+                    decimal i;
+                    if (decimal.TryParse(txtCredit.Text, out i))
                     {
                         //輸入是數字
                         if (i < 0)
